backend/requirements/base.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
django==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Search
elasticsearch==8.11.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Task Queue & Messaging
celery==5.3.6
kombu==5.3.5
redis==5.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
